// packages/shared/src/separation/mvsep-catalog.generated.ts
export const MVSEP_CATALOG_DATE = '2026-10-02'

export const MVSEP_ALGORITHMS: Record<number, MvsepAlgorithm> = {
  "0": {
    "name": "spleeter",
    "priceCoefficient": 1,
    "outputs": {
      "vocals": "Vocals",
      "accompaniment": "Accompaniment"
    },
    "fields": {
      "add_opt1": {
        "serverKey": "model_type",
        "label": "Model type",
        "default": "0",
        "options": {
          "0": "2 stems (vocals, music)",
          "1": "4 stems (vocals, drums, bass, other)",
          "2": "5 stems (vocals, drums, bass, piano, other)"
        }
      }
    }
  },
// ...
}